requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
certifi==2024.12.14
enum34==1.1.10
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
netifaces==0.11.0
propcache==0.2.1
psutil==6.1.1
pydantic==2.10.5
pydantic_core==2.27.2
pyspectator==1.2.2
python-dotenv==1.0.1
pyvalid==1.0.4
six==1.17.0
typing_extensions==4.12.2
yarl==1.18.3
